fix: Move files, fix type hints
